Refactor ApiData mapping and share the map code with Form.
